requirements.txt
altgraph==0.17.2
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
cryptography==37.0.2
dill==0.3.5.1
idna==3.3
macholib==1.16
mpmath==1.2.1
ntlm-auth==1.5.0
numpy==1.23.0
pbr==5.9.0
Pillow==9.2.0
ply==3.11
psutil==5.9.1
pycparser==2.21
pygame==2.1.2
pyinstaller-hooks-contrib==2022.8
python-dateutil==2.8.2
qiskit==0.37.0
qiskit-aer==0.10.4
qiskit-ibmq-provider==0.19.2
qiskit-terra==0.21.0
requests==2.28.1
requests-ntlm==1.1.0
retworkx==0.11.0
scipy==1.8.1
six==1.16.0
stevedore==3.5.0
symengine==0.9.2
sympy==1.10.1
tweedledum==1.1.1
urllib3==1.26.9
websocket-client==1.3.3
websockets==10.3
